requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
PyYAML
gunicorn
psycopg[binary,pool]
